requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django==5.1.3
django-cors-headers==4.6.0
django-userforeignkey==0.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
psycopg2==2.9.10
PyJWT==2.10.0
pymongo==4.10.1
sqlparse==0.5.2
